Pull out some of the mint logic for #22
